requirements.txt
asgiref==3.7.2
boto3==1.27.0
botocore==1.30.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.3
oauthlib==3.2.2
Pillow==10.0.0
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
